pyproject.toml
[project]
name = "h3ntai-hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.2",
    "natsort>=8.4.0",
    "patool>=4.0.1",
    "playwright>=1.52.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
]
